pyproject.toml
[project]
name = "project1-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "cairosvg>=2.8.2",
    "fastapi>=0.115.12",
    "google-generativeai>=0.8.5",
    "html2text>=2025.4.15",
    "langchain>=0.3.25",
    "markdownify>=1.1.0",
    "nltk>=3.9.1",
    "numpy>=2.3.0",
    "pillow>=11.2.1",
    "playwright>=1.52.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "sentence-transformers>=4.1.0",
    "tiktoken>=0.9.0",
    "uvicorn>=0.34.3",
]
